# contracts/api/inventory-service.openapi.yaml
openapi: 3.1.0
info:
  title: ERClave Inventory Service API
  version: 0.2.0
  summary: Warehouse, item, movement, balance, kardex and reservation APIs.
servers:
  - url: https://api.eslaclave.com
security:
  - bearerAuth: []
...
paths:
  /v1/inventory/reports/{report_code}/export:
    get:
      tags: [Reports]
      operationId: exportInventoryStandardReport
      summary: Export a fixed Inventory report as UTF-8 CSV.
      x-required-module: inventory
      x-permissions: [inventory.warehouse.read, inventory.item.read, inventory.balance.read, inventory.kardex.read]
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - { name: report_code, in: path, required: true, schema: { type: string } }
        - { name: lang, in: query, schema: { type: string, enum: [es, en], default: es } }
        - { name: date_from, in: query, schema: { type: string, format: date } }
        - { name: date_to, in: query, schema: { type: string, format: date } }
        - { name: status, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string, maxLength: 120 } }
        - { name: warehouse_id, in: query, schema: { type: string } }
        - { name: item_id, in: query, schema: { type: string } }
        - { name: movement_type, in: query, schema: { type: string } }
        - { name: condition, in: query, schema: { type: string } }
      responses:
        '200': { description: CSV export. }
        '204': { description: No rows matched the filters. }
  /v1/inventory/warehouses:
    get:
      tags: [Warehouses]
      operationId: listWarehouses
      summary: List warehouses.
      x-required-module: inventory
      x-permissions: [inventory.warehouse.read, maintenance.material_request.read, maintenance.material_request.create]
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: Warehouses.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseListResponse'
    post:
      tags: [Warehouses]
      operationId: createWarehouse
      summary: Create warehouse.
      x-required-module: inventory
      x-permissions: [inventory.warehouse.create]
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseCreateRequest'
      responses:
        '201':
          description: Warehouse created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseResponse'
  /v1/inventory/warehouses/{id}:
    get:
      tags: [Warehouses]
      operationId: getWarehouse
      summary: Get one tenant warehouse for Inventory or Purchasing validation.
      x-required-module: inventory
      x-permissions: [inventory.warehouse.read, purchasing.receipt.create, maintenance.material_request.create]
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Warehouse.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseResponse'
    patch:
      tags: [Warehouses]
      operationId: updateWarehouse
      summary: Update warehouse.
      x-required-module: inventory
      x-permissions: [inventory.warehouse.update]
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseUpdateRequest'
      responses:
        '200':
          description: Warehouse updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseResponse'
...
  /v1/inventory/items:
    get:
      tags: [Items]
      operationId: listInventoryItems
      summary: Search inventory items.
      x-required-module: inventory
      x-permissions: [inventory.item.read, production.product_service.create, production.product_service.update, sales.order.fulfill, maintenance.material_request.read, maintenance.material_request.create, purchasing.requisition.create, purchasing.requisition.update, purchasing.order.create, purchasing.order.update, purchasing.order.issue]
      x-permission-mode: any
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Status'
        - name: use_in_recipe
          in: query
          description: Return only items explicitly eligible or ineligible for production recipes.
          schema:
            type: boolean
        - name: type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemListResponse'
    post:
      tags: [Items]
      operationId: createInventoryItem
      summary: Create inventory item.
      x-required-module: inventory
      x-permissions: [inventory.item.create]
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreateRequest'
      responses:
        '201':
          description: Item created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
  /v1/inventory/items/{id}:
    get:
      tags: [Items]
      operationId: getInventoryItem
      summary: Get inventory item.
      x-required-module: inventory
      x-permissions: [inventory.item.read, production.product_service.create, production.product_service.update, sales.order.fulfill, maintenance.material_request.create, purchasing.requisition.create, purchasing.requisition.update, purchasing.order.create, purchasing.order.update, purchasing.order.issue]
      x-permission-mode: any
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Item detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
    patch:
      tags: [Items]
      operationId: updateInventoryItem
      summary: Update inventory item.
      x-required-module: inventory
      x-permissions: [inventory.item.update]
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdateRequest'
      responses:
        '200':
          description: Item updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
...
  /v1/inventory/balances:
    get:
      tags: [Balances]
      operationId: listInventoryBalances
      summary: List calculated inventory balances.
      x-required-module: inventory
      x-permissions: [inventory.balance.read]
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Search'
        - name: inventory_item_id
          in: query
          schema:
            type: string
        - name: warehouse_id
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema: { type: string }
        - name: item_type
          in: query
          schema: { type: string }
        - name: item_status
          in: query
          schema: { type: string, enum: [active, inactive, blocked] }
        - name: inventory_policy
          in: query
          schema: { type: string }
        - name: unit
          in: query
          schema: { type: string }
        - name: stock_status
          in: query
          schema: { type: string, enum: [negative, out_of_stock, below_minimum, normal, above_maximum] }
        - name: sort
          in: query
          schema: { type: string, enum: [item_code, item_name, on_hand_asc, on_hand_desc], default: item_code }
      responses:
        '200':
          description: Balances.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceListResponse'
  /v1/inventory/kardex:
    get:
      tags: [Balances]
      operationId: listKardex
      summary: List kardex entries calculated from movements.
      x-required-module: inventory
      x-permissions: [inventory.kardex.read]
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: inventory_item_id
          in: query
          schema:
            type: string
        - name: warehouse_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Kardex.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovementListResponse'
...
  /v1/inventory/purchase-receipts:
    post:
      tags: [Movements]
      operationId: createInventoryPurchaseReceiptMovement
      summary: Register a purchase-order entry requested by Purchasing.
      x-required-module: inventory
      x-permissions: [purchasing.receipt.create]
      parameters:
        - $ref: '#/components/parameters/TenantHeader'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovementCreateRequest'
      responses:
        '201':
          description: Purchase receipt movement recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovementResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantHeader:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        maximum: 200
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
    Search:
      name: q
      in: query
      schema:
        type: string
    Status:
      name: status
      in: query
      schema:
        type: string
  schemas:
...
    WarehouseResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Warehouse'
    WarehouseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Warehouse'
    WarehouseCreateRequest:
      type: object
      required: [code, name, type, business_center, location, owner]
      properties:
        code:
          type: string
        name:
          type: string
        type:
          type: string
        business_center: { type: string }
        location: { type: string }
        owner: { type: string }
        capacity: { type: [string, 'null'] }
        inventory_policy: { type: string, default: standard }
        zone: { type: [string, 'null'] }
        aisle: { type: [string, 'null'] }
        rack: { type: [string, 'null'] }
        level: { type: [string, 'null'] }
        position: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
...
    ItemListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        page:
          $ref: '#/components/schemas/Page'
...